app/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.4
djangorestframework==3.14.0
dnspython==2.2.1
flower==1.2.0
gunicorn==20.1.0
humanize==4.4.0
kombu==5.2.4
numpy==1.23.5
pandas==1.5.2
prometheus-client==0.15.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pymongo==4.3.3
python-dateutil==2.8.2
pytz==2022.6
redis==4.4.0
six==1.16.0
sqlparse==0.4.3
tornado==6.2
vine==5.0.0
wcwidth==0.2.5
